Porting fcn and lraspp
